proof refactor, include every field in the calculation
Before this commit, we relied on a separate struct which contained every DidDocument field except Proof, we then JSON-marshal it and use the result to calculate a matching Proof field: if what we calculate doesn't match what the user provides, the proof signature fails and the according DidDocument doesn't get added to the chain.

We now define DidDocumentUnsigned as a type alias to DidDocument just to make it clear we're dealing with unsigned data.

In the VerifyProof function we make sure to explicitly zero out the Proof field, effectively creating a deep copy of the original DidDocument containing only the data we must check, making future updates to the DidDocument struct as painless as possible.