first implementation of per-service signature cost
This commit introduces a first draft implementation of a document sharing scheme which is fully aware of the potential cost associated with document signature.

Each Service provider (hence, each DDO entity) can list a set of Signature prices associated with a Signature scheme.

The chain will send the amount of tokens associated with the signature scheme to the Signature provider itself, associating the message sharing with the message signature request and compensation.

This system supports free signatures schemes (e.g. "dumb" signature might be provided for free) represented by Signature schemes associated with a nil price, as well as price-free signature providers.

Since this is a blockchain-level protocol change, the Signature providers will have to implement application-level policies for failed signatures (for example, reimburse the user if, for any reason, the process result in a error).

x/docs tests are failing right now, I'm in the process of making everything work again.